Move models package into subpackage of storage
