+++ server/internal/storage/models/artifact.go
